// lib/types.ts
export type Property = { 
  id: string; 
  title: string; 
  location: string; 
  imageAlt: string; 
  beds: number; 
  baths: number; 
  guests: number; 
  priceFrom: number; 
  href: string 
};

export type ExperienceCard = { 
  title: string; 
  blurb: string; 
  imageAlt: string; 
  href: string 
};